Dynamically update description size
Found the font.GetStringSize method, which returns the size of a box needed to fit a given text of that font size. This allows me to update the description line by line, finding the amount of lines that fit within the box and adjusting the height by that  many lines.